+++ OneBlink.SDK/models/BPoint/BPointBankAccountDetails.cs
@@ -1,0 +1,18 @@
+namespace OneBlink.SDK.Model
+{
+    public class BPointBankAccountDetails
+    {
+        public string AccountName
+        {
+            get; set;
+        }
+        public string BSBNumber
+        {
+            get; set;
+        }
+        public string TruncatedAccountNumber
+        {
+            get; set;
+        }
+    }
+}