+++ OneBlink.SDK/models/BPoint/BPointCVNResult.cs
@@ -1,0 +1,11 @@
+namespace OneBlink.SDK.Model
+{
+    public class BPointCVNResult
+    {
+        public string CRNResultCode
+        {
+            get; set;
+        }
+
+    }
+}