+++ OneBlink.SDK/models/BPoint/BPointReDResponse.cs
@@ -1,0 +1,38 @@
+namespace OneBlink.SDK.Model
+{
+    public class BPointReDResponse
+    {
+        public string REQ_ID
+        {
+            get; set;
+        }
+        public string ORD_ID
+        {
+            get; set;
+        }
+        public string STAT_CD
+        {
+            get; set;
+        }
+        public string FRAUD_STAT_CD
+        {
+            get; set;
+        }
+        public string FRAUD_RSP_CD
+        {
+            get; set;
+        }
+        public string FRAUD_REC_ID
+        {
+            get; set;
+        }
+        public string FRAUD_NEURAL
+        {
+            get; set;
+        }
+        public string FRAUD_RFC
+        {
+            get; set;
+        }
+    }
+}