+++ OneBlink.SDK/models/BPoint/BPointThreeDSResponse.cs
@@ -1,0 +1,38 @@
+namespace OneBlink.SDK.Model
+{
+    public class BPointThreeDSResponse
+    {
+        public string Eci
+        {
+            get; set;
+        }
+        public string Enrolled
+        {
+            get; set;
+        }
+        public string Status
+        {
+            get; set;
+        }
+        public string VerifySecurityLevel
+        {
+            get; set;
+        }
+        public string VerifyStatus
+        {
+            get; set;
+        }
+        public string VerifyToken
+        {
+            get; set;
+        }
+        public string VerifyType
+        {
+            get; set;
+        }
+        public string Xid
+        {
+            get; set;
+        }
+    }
+}