+++ OneBlink.SDK/models/FormSubmissionMetaResponse.cs
@@ -6,22 +6,23 @@ namespace OneBlink.SDK.Model
 {
     public class FormSubmissionMetadataResponse
     {
-        [JsonProperty]
         public FormSubmissionMetadata formSubmissionMeta
         {
-            get; internal set;
+            get; set;
         }
 
-        [JsonProperty]
         public FormApprovalFlowInstance formApprovalFlowInstance
         {
-            get; internal set;
+            get; set;
         }
 
-        [JsonProperty]
         public IEnumerable<FormSubmissionApproval> formSubmissionApprovals
         {
-            get; internal set;
+            get; set;
+        }
+        public FormSubmissionPayment formSubmissionPayment
+        {
+            get; set;
         }
     }
 }